Extract StringConverter to a new class
